backend/requirements.txt
flask==2.3.2
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
psycopg2==2.9.6
python-dotenv==1.0.0
rich
Flask-Migrate